Refactor test_render_script_integration and feature-gate
