+++ crates/pluot/Cargo.toml
@@ -28,14 +28,20 @@ wasm-opt = ['-Oz']
 
 [features]
 default = []
+# Embed URW Core 35 fonts via include_bytes! for plain-Rust use (enables PDF Base-14 font tests)
+embed_fonts = ["pluot_core/embed_fonts"]
 # Define features to toggle which dependency is active
 python = ["pluot_core/python", "pyo3"]
 # R bindings feature
 rlang = ["pluot_core/rlang"]
 # Define a feature for running unit tests in GPU-less environments
 lacks_gpu = ["pluot_core/lacks_gpu"]
-# Embed URW Core 35 fonts via include_bytes! for plain-Rust use (enables PDF Base-14 font tests)
-embed_fonts = ["pluot_core/embed_fonts"]
+# Define a feature for running render_script tests in CI envs that lack Python envs.
+lacks_python_env = []
+# Define a feature for running render_script tests in CI envs that lack Python envs.
+lacks_rlang_env = []
+
+
 
 
 [dependencies]
