+++ crates/pluot/tests/test_utils/mod.rs
@@ -1,2 +1,4 @@
+mod render_script_integration;
 mod snapshot_utils;
+pub use render_script_integration::*;
 pub use snapshot_utils::*;
